Remove legacy library code
removed ununsed library routes and leaving only library-dataset-view.js behind
